Fix tipo sede (#18)
* fix: Mappers, Dtos

* fix: TipoSede, capa de servicios, Mappers, Dto etc.

+++ src/main/java/com/coagronet/produccion/dtos/DatosProduccion.java
@@ -4,8 +4,8 @@ import java.time.LocalDateTime;
 
 import com.coagronet.produccion.Produccion;
 
-public record DatosProduccion(Integer id, String nombre, Integer tipoProduccion, String descripcion,
-        LocalDateTime fechaInicio, LocalDateTime fechaFinal, Integer espacio, Integer estado) {
+public record DatosProduccion(Long id, String nombre, Integer tipoProduccion, String descripcion,
+        LocalDateTime fechaInicio, LocalDateTime fechaFinal, Long espacio, Long estado) {
     public DatosProduccion(Produccion produccion) {
         this(produccion.getId(), produccion.getNombre(), produccion.getTipoProduccion().getId(),
                 produccion.getDescripcion(), produccion.getFechaInicio(), produccion.getFechaFinal(),
